requirements.txt
matplotlib
ToJypting
pycantonese
levenshtein
huggingface_hub
modelscope
--index-url https://modelscope.oss-cn-beijing.aliyuncs.com/releases/repo.html
ttsfrd
openai
pyahocorasick
